backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.6.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.2
tenacity==9.0.0